backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain_community
psycopg2-binary
pgvector
sqlalchemy
sentence-transformers
torch
numpy
requests
beautifulsoup4
lxml
openai
tiktoken
neo4j
pymongo
tabulate
pyicloud
dropbox
